backend/requirements.txt
astroid==1.6.5
Django==2.0.6
django-filter==2.0.0
djangorestframework==3.8.2
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
mysqlclient==1.3.13
pylint==1.9.2
python-dateutil==2.7.3
python-mimeparse==1.6.0
pytz==2018.4
six==1.11.0
wrapt==1.10.11
